[target-core] target_s3_jsonl renamed as target_s3_json
