+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = target-s3-jsonl
-version = attr: target_s3_jsonl.__version__
+version = attr: target_s3_json.__version__
 description = Singer.io target for writing JSON Line files and upload to S3
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -26,7 +26,7 @@ project_urls =
 package_dir =
     = src
 packages = find:
-py_modules = target_s3_jsonl
+py_modules = target_s3_json
 python_requires = >=3.9
 # install_requires =
 #     jsonschema==4.9.1
@@ -36,7 +36,7 @@ include_package_data = True
 platforms = any
 
 [options.package_data]
-target_s3_jsonl = logging.conf
+target_s3_json = logging.conf
 
 [options.packages.find]
 where = src
@@ -45,7 +45,7 @@ exclude =
 
 [options.entry_points]
 console_scripts =
-    target-s3-jsonl = target_s3_jsonl:main
+    target-s3-jsonl = target_s3_json:main
 
 [options.extras_require]
 test =
@@ -63,7 +63,7 @@ deploy = twine
 doc = sphinx-rtd-theme
 
 [tool:pytest]
-addopts = -v --cov=target_s3_jsonl --cov-fail-under 95 --cov-report xml --cov-report term --cov-report html:htmlcov --doctest-modules
+addopts = -v --cov=target_s3_json --cov-fail-under 95 --cov-report xml --cov-report term --cov-report html:htmlcov --doctest-modules
 testpaths = tests
 asyncio_mode = auto
